Move directory only used by src files to src directory
